change month type i32 -> u32

+++ src/date.rs
@@ -2,7 +2,7 @@ fn is_leap_year(year: i32) -> bool {
     year % 4 == 0 && (year % 400 == 0 || year % 100 != 0)
 }
 
-pub fn days_in_month(month: u32, year: i32) -> i64 {
+pub fn days_in_month(month: u32, year: i32) -> u32 {
     match month {
         4 | 6 | 9 | 11 => 30,
         2 if is_leap_year(year) => 29,
@@ -11,7 +11,7 @@ pub fn days_in_month(month: u32, year: i32) -> i64 {
     }
 }
 
-pub fn days_in_year(year: i32) -> i64 {
+pub fn days_in_year(year: i32) -> u32 {
     if is_leap_year(year) {
         366
     } else {
